stateclass refactor
