+++ mesop/BUILD
@@ -53,6 +53,7 @@ py_library(
         "//mesop/commands",
         "//mesop/features",
         "//mesop/server",
+        "//mesop/stateclass",
     ],
 )
