+++ mesop/stateclass/BUILD
@@ -1,0 +1,13 @@
+load("//build_defs:defaults.bzl", "py_library")
+
+package(
+    default_visibility = ["//build_defs:mesop_internal"],
+)
+
+py_library(
+    name = "stateclass",
+    srcs = glob(["*.py"]),
+    deps = [
+        "//mesop/runtime",
+    ],
+)
